feat: add add todo hook
