+++ homeworks/anastasia-kladova/20-react-api/src/components/Form/Form.jsx
@@ -3,9 +3,7 @@ import Input from '../Input/Input';
 import Button from '../Buttons/Button/Button';
 import {useContext} from 'react';
 import {Context} from '../../contexts/AppContext/AppContext';
-import {useValidateTodo} from '../../hooks/validateTodoHook';
-import {ACTION_TYPES} from '../../state/actionTypes';
-import {useAddTodo} from '../../hooks/deleteTodoHook';
+import {useAddTodo} from '../../hooks/addTodoHook';
 
 const Form = () => {
   const {
@@ -18,27 +16,10 @@ const Form = () => {
     setIsNewTodoValid,
   } = useContext(Context);
 
-  const newTodo = currentInputText.trim();
-  const {currentError} = useValidateTodo(newTodo, state.todos);
+  const {addNewTodo} = useAddTodo();
 
   const handleCurrentInputValue = e => {
     setCurrentInputText(e.target.value);
-  };
-
-  const addNewTodo = e => {
-    e.preventDefault();
-
-    if (!currentError) {
-      dispatch({type: ACTION_TYPES.ADD_NEWTODO, payload: {newTodo: newTodo}});
-      setCurrentInputText('');
-      inputRef.current.focus();
-      setIsNewTodoValid(false);
-      setErrorMessage('');
-    } else {
-      setIsNewTodoValid(true);
-      setErrorMessage(currentError);
-      setCurrentInputText('');
-    }
   };
 
   return (
